Revert "Refactor radioButtonList (#2642)"
This reverts commit c8b7c6f30a5c0e9f1e061614b128cbbe29369836.
