+++ src/components/RadioButtonList/index.tsx
@@ -2,4 +2,4 @@
 // Copyright (C) 2021 - 2023 Gemeente Amsterdam
 export { default } from './RadioButtonList'
 export type { RadioButtonListProps } from './RadioButtonList'
-export type { RadioButtonOption } from './RadioButtonOption'
+export type { RadioButtonOption } from './RadioButtonList'
